Fixes after react-northstar merge
