+++ packages/fluentui/react-northstar/src/themes/teams/components/Card/cardHeaderStyles.ts
@@ -8,9 +8,9 @@ const cardHeaderStyles: ComponentSlotStylesPrepared<CardHeaderStylesProps, CardV
       display: 'flex',
       flexDirection: 'column',
       margin: v.headerMargin,
-      ...(p.fitted && { margin: 0 })
+      ...(p.fitted && { margin: 0 }),
     };
-  }
+  },
 };
 
 export default cardHeaderStyles;
