+++ packages/fluentui/react-northstar/src/themes/teams/components/Card/cardHeaderVariables.ts
