+++ packages/fluentui/react-northstar/src/themes/teams/components/Card/cardTopControlsStyles.ts
@@ -7,9 +7,9 @@ const cardTopControlsStyles: ComponentSlotStylesPrepared<CardTopControlsStylesPr
     return {
       position: 'absolute',
       top: v.topControlsTop,
-      right: v.topControlsRight
+      right: v.topControlsRight,
     };
-  }
+  },
 };
 
 export default cardTopControlsStyles;
